requirements.txt
selenium
bs4
conan
